requirements.txt
alembic==1.7.3
amqp==5.0.9
asgiref==3.4.1
attrs==21.2.0
bcrypt==3.2.0
billiard==3.6.4.0
boto3==1.18.64
botocore==1.21.64
celery==5.2.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
Deprecated==1.2.13
Django==4.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
et-xmlfile==1.1.0
fastapi==0.68.1
greenlet==1.1.1
h11==0.12.0
httptools==0.2.0
idna==3.2
image==1.5.33
iniconfig==1.1.1
jmespath==0.10.0
kombu==5.2.2
Mako==1.1.5
MarkupSafe==2.0.1
openpyxl==3.0.9
packaging==21.0
passlib==1.7.4
Pillow==9.0.0
pluggy==1.0.0
prompt-toolkit==3.0.24
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyheif==0.6.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.5
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
six==1.16.0
SQLAlchemy==1.4.25
sqlalchemy2-stubs==0.0.2a15
sqlparse==0.4.2
starlette==0.14.2
starlette-context==0.3.3
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.0
wrapt==1.13.3
